requirements.txt
requests==2.32.3
pandas==2.2.3
pyarrow==17.0.0
python-dotenv==1.0.1
tenacity==9.0.0
pytest==8.3.3
responses==0.25.3